Versioning the API
Why: So users can continue to use the api even when new version is
released

This change addresses the need by:
*Creating api/v1/ directories under the controller directory
*Updating route.rb by namespacing the resources under api and scoping to
module v1(setting defualt to true )
*Setting the module constraint to ApiConstraints
*Creating lib/api_constraints.rb to manage the constraints
*Updating the auth_controller.rb, bucketlists_controller.rb,
items_controller.rb and users_controller.rb to reflect under the api/v1
module.

[Finishes #111680752]
