+++ config/routes.rb
@@ -1,14 +1,13 @@
 Rails.application.routes.draw do
-  resources :bucketlists do
-    resources :items, only: [:create, :update, :destroy]
+  namespace :api, defaults: {format: 'json'}, constraints: {}, path: '/' do
+    scope module: :v1,
+                  constraints: ApiConstraints.new(version: 1, default: true) do
+      resources :bucketlists do
+        resources :items, only: [:create, :update, :destroy]
+      end
+      post 'auth/login'
+      get 'auth/logout'
+      resources :users, :only => ["create"]
+    end
   end
-  post 'auth/login'
-
-  get 'auth/logout'
-
-  resources :users, :only => ["create"]
-  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
-
-  # Serve websocket cable requests in-process
-  # mount ActionCable.server => '/cable'
 end
