+++ lib/api_constraints.rb
@@ -1,0 +1,10 @@
+class ApiConstraints
+  def initialize(options)
+    @version = options[:version]
+    @default = options[:default]
+  end
+
+  def matches?(req)
+    @default || req.headers['Accept'].include?("application/vnd.matbucketlist.v#{@version}")
+  end
+end
